requirements.txt
absl-py==1.4.0
alembic==1.9.1
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
autopage==0.5.1
avalanche-lib==0.3.1
backcall==0.2.0
beautifulsoup4==4.11.1
bottle==0.12.23
cachetools==4.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.1.3
cliff==4.1.0
cmaes==0.9.1
cmd2==2.4.2
colorama==0.4.6
colorlog==6.7.0
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.15
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
dm-tree==0.1.8
docker-pycreds==0.4.0
entrypoints==0.4
etils==1.1.0
executing==1.2.0
filelock==3.8.2
flatbuffers==23.3.3
fonttools==4.38.0
gast==0.4.0
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.58.0
googledrivedownloader==0.4
GPUtil==1.4.0
greenlet==2.0.1
grpcio==1.27.2
gunicorn==20.1.0
h5py==3.8.0
idna==2.10
importlib-metadata==4.13.0
importlib-resources==5.10.2
install==1.3.5
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==7.4.9
jupyter_core==5.1.4
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Mako==1.2.4
Markdown==3.2.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mpmath==1.2.1
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
oauthlib==3.2.2
opt-einsum==3.3.0
optuna==3.0.5
optuna-dashboard==0.8.1
optuna-fast-fanova==0.0.4
packaging==22.0
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
platformdirs==2.6.2
prettytable==3.6.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyperclip==1.8.2
pyreadline3==3.4.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
pytorchcv==0.0.67
pytz==2022.7.1
pywin32==306
PyYAML==6.0
pyzmq==25.0.0
quadprog==0.1.7
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.24.1
scipy==1.8.1
seaborn==0.12.2
sentry-sdk==1.12.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.14.0
sklearn==0.0
smmap==3.0.4
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
stack-data==0.6.2
stevedore==4.1.1
sympy==1.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-datasets==4.8.3
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.12.0
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchmetrics==0.11.0
torchshow==0.5.0
torchvision==0.15.2+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.16
wandb==0.13.7
wcwidth==0.2.5
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.11.0
